extendedPin functionality

+++ testData/generator/LeftAssociative.expected.java
@@ -47,17 +47,17 @@ public class LeftAssociative implements PsiParser {
   public static boolean alias_definition(PsiBuilder builder_, int level_) {
     if (!recursion_guard_(builder_, level_, "alias_definition")) return false;
     boolean result_ = false;
-    final Marker left_marker_ = ((Marker)builder_.getLatestDoneMarker()).precede();
+    final Marker left_marker_ = (Marker)builder_.getLatestDoneMarker();
+    if (!invalid_left_marker_guard_(builder_, left_marker_, "alias_definition")) return false;
     final Marker marker_ = builder_.mark();
     result_ = alias_definition_0(builder_, level_ + 1);
     result_ = result_ && consumeToken(builder_, ID);
     if (result_) {
       marker_.drop();
-      left_marker_.done(ALIAS_DEFINITION);
+      left_marker_.precede().done(ALIAS_DEFINITION);
     }
     else {
       marker_.rollbackTo();
-      left_marker_.drop();
     }
     return result_;
   }
@@ -81,17 +81,17 @@ public class LeftAssociative implements PsiParser {
   private static boolean alias_definition2_0(PsiBuilder builder_, int level_) {
     if (!recursion_guard_(builder_, level_, "alias_definition2_0")) return false;
     boolean result_ = false;
-    final Marker left_marker_ = ((Marker)builder_.getLatestDoneMarker()).precede();
+    final Marker left_marker_ = (Marker)builder_.getLatestDoneMarker();
+    if (!invalid_left_marker_guard_(builder_, left_marker_, "alias_definition2_0")) return false;
     final Marker marker_ = builder_.mark();
     result_ = alias_definition2_0_0(builder_, level_ + 1);
     result_ = result_ && consumeToken(builder_, ID);
     if (result_) {
       marker_.drop();
-      left_marker_.done(ALIAS_DEFINITION2);
+      left_marker_.precede().done(ALIAS_DEFINITION2);
     }
     else {
       marker_.rollbackTo();
-      left_marker_.drop();
     }
     return result_;
   }
@@ -159,6 +159,7 @@ public class LeftAssociative implements PsiParser {
     if (!recursion_guard_(builder_, level_, "leech")) return false;
     boolean result_ = false;
     final Marker left_marker_ = (Marker)builder_.getLatestDoneMarker();
+    if (!invalid_left_marker_guard_(builder_, left_marker_, "leech")) return false;
     final Marker marker_ = builder_.mark();
     result_ = consumeToken(builder_, ID);
     if (result_) {
@@ -179,6 +180,7 @@ public class LeftAssociative implements PsiParser {
     if (!recursion_guard_(builder_, level_, "leech2")) return false;
     boolean result_ = false;
     final Marker left_marker_ = (Marker)builder_.getLatestDoneMarker();
+    if (!invalid_left_marker_guard_(builder_, left_marker_, "leech2")) return false;
     final Marker marker_ = builder_.mark();
     result_ = consumeToken(builder_, ID);
     if (!result_) {
